+++ src/common/StockSymbolList/index.js
