--- src/components/SearchStock/SearchStock.js
@@ -1,11 +1,0 @@
-import React from "react";
-
-const SearchStock = () => {
-  return (
-    <div>
-      <h1>Hello Stocks!</h1>
-    </div>
-  );
-};
-
-export default SearchStock;
