--- src/components/StockSheetItem/StockSheetItem.module.scss
@@ -1,23 +1,0 @@
-.root {
-  border-bottom: 1px solid rgb(143, 143, 143);
-
-  .price {
-    color: rgb(112, 112, 112);
-  }
-  .plus_percent {
-    background-color: rgb(0, 187, 0);
-    height: 20px;
-    width: 50px;
-    padding: 2px 0px 0px 2px;
-    border-radius: 4px;
-    color: white;
-  }
-  .minus_percent {
-    background-color: rgb(253, 7, 7);
-    height: 20px;
-    width: 50px;
-    padding: 2px 0px 0px 2px;
-    border-radius: 4px;
-    color: white;
-  }
-}
