--- src/components/stock/StockBordItem.module.css
@@ -1,19 +1,0 @@
-.content {
-  width: 150px;
-  height: 50px;
-  background-color: rgb(56, 56, 56);
-  padding-top: 15px;
-  color: aliceblue;
-}
-
-.red {
-  background-color: rgb(235, 0, 0);
-  border-radius: 4px;
-  padding: 0px 10px;
-}
-
-.green {
-  background-color: rgb(0, 187, 0);
-  border-radius: 4px;
-  padding: 0px 10px;
-}
