+++ src/containers/HomeChart/Chart.module.scss
@@ -2,9 +2,6 @@
   .stock_info {
     width: 230px;
 
-    .symbol {
-      border-bottom: 1px solid black;
-    }
     .stock_info_right {
       .current_price {
         color: rgb(88, 88, 88);
